Cleanup of ComposeActivity
